requirements.txt
# Procesamiento de datos
pandas==2.0.3
numpy==1.24.3
pymongo==4.4.1
python-dotenv==1.0.0

# Visualización
plotly==5.15.0
dash==2.11.1
dash-bootstrap-components==1.4.1

# Análisis estadístico
scipy==1.11.1
scikit-learn==1.3.0

# Utilidades
requests==2.31.0

# Despliegue
gunicorn==20.1.0
